+++ CHANGELOG.md
@@ -12,9 +12,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Updated sqlparser version
 - Deactivate graphviz display by default
 - Deactivate multiplicity testing by default
-- Fixed Index trait for `data_type::Value` and `DataType`
+- Improved Index trait for `data_type::Value` and `DataType`[MR94](https://github.com/Qrlew/qrlew/pull/94)
 ### Added
-- Implemented `hierarchy` method for `data_type::Value` and `DataType`
+- Implemented `hierarchy` method for `data_type::Value` and `DataType`[MR94](https://github.com/Qrlew/qrlew/pull/94)
 
 ## [0.2.1] - 2023-07-26
 ### Added